NTP: Replace Chromium's NTP with the Rebel RemoteNTP
The RemoteNTP is a React application to serve as the NTP for each
platform (including mobile):

https://github.com/RebelBrowser/remote_ntp

This includes all hooks from the renderer process to the browser process
to embed APIs needed to create an NTP (most visited tiles, theme, etc.).

This also includes a DEPS hook to sync the RemoteNTP repo and build it
for a local copy of the NTP. Under normal circumstances, the NTP will
always be loaded remotely. But if the browser is offline or if any other
network error occurs, chrome://remote-ntp-offline will be loaded instead
using this local copy.