Branding: Remove lxml requirement from create_branded_grd.py
This was handy when this script was originally written to generate
source files meant to live in-tree. It would allow diff'ing between the
generated files and the orginal source file without much noise. But now
the generated files are placed in out/Default/gen, and not checked into
the git repository.