Branding: Display chrome:// URLs with a Rebel-branded schema
This essentially creates an alias from the Rebel-branded schema (e.g.
rebel://) to the Chrome schema (chrome://). The Chrome schema itself is
just too pervasive to replace wholesale, so treating the Rebel-branded
schema as an alias avoids breaking the browser; we just need to ensure
the places it is displayed to the user becomes branded.