Prism: Add a few log points for tracing the prefetch of Prism hints
This adds log points to the OGS log to indicate when a prefetch request
is queued (i.e. waiting for the prefetch manager to actually issue the
request), when the request is issued, and when the response is received.

This requires viasat://flags/#optimization-guide-debug-logs to be turned
on. The OGS log is available at viasat://optimization-guide-internals/.