Prism: Do not override exisiting paramters with --enable-features
We currently enable Prism-related Loading Predictor features by setting
the --enable-features command line flag. However, if we set that command
like flag ourselves when starting VB, we would ignore any value passed
in and replace it with our own (empty) value. For example, if this flag
was set on the command line as:

  --enable-features="LoadingPredictorPrefetch:subresource_type/css"

We would ignore the "subresource_type" parameter and just set:

  --enable-features="LoadingPredictorPrefetch"

This change makes us respect anything on the command line already.