Prism: Set the Accept header on prefetch requests
If the content server responds with a "Vary: Accept" header, the cached
prefetch response is unlikely to be used without this request header.
For example, organic requests through Blink will have an Accept header
value of:

    image/avif,image/webp,image/apng,image/svg+xml,image/*,*/*;q=0.8

So if the response headers have "Vary: Accept", and the prefetch request
did not have this Accept header, the cached response is not used due to
a Vary mismatch.